Handle no-valid characters in sanitization
