+++ discovery-provider/src/api/v1/utils/resolve_url.py
@@ -1,9 +1,9 @@
 import re
+from src.utils.helpers import encode_int_id
 from flask.helpers import url_for
 from urllib.parse import urlparse
 from src.models import User
 from src.api.v1 import api as api_v1
-from src.api.v1.helpers import encode_int_id
 from src.api.v1.tracks import ns as tracks_ns
 from src.api.v1.users import ns as users_ns
 from src.api.v1.playlists import ns as playlists_ns
